Changed Mondrian version to 3.4.0-SNAPSHOT and updated source file to handle move of Metadata Location in platform
