+++ build.properties
@@ -21,7 +21,7 @@ dependency.reporting-library.group=pentaho-library
 dependency.kettle.revision=TRUNK-SNAPSHOT
 dependency.libbase.revision=1.2-SNAPSHOT
 dependency.libformula.revision=1.2-SNAPSHOT
-dependency.mondrian.revision=3.3-SNAPSHOT
+dependency.mondrian.revision=3.4.0-SNAPSHOT
 dependency.pentaho-actionsequence-dom.revision=TRUNK-SNAPSHOT
 dependency.pentaho-connections.revision=TRUNK-SNAPSHOT
 dependency.pentaho-cwm.revision=TRUNK-SNAPSHOT
