+++ test-res/solution1/system/data/sampledata.properties
@@ -1,5 +1,5 @@
 #HSQL Database Engine 1.8.0.5
-#Wed Sep 21 19:26:10 EDT 2011
+#Wed Feb 15 09:56:00 EST 2012
 hsqldb.script_format=0
 runtime.gc_interval=0
 sql.enforce_strict_size=false
